requirements.txt
absl-py==0.7.1
alabaster==0.7.12
alembic==1.7.7
allennlp==0.8.3
altair==1.2.0
ansiwrap==0.8.4
appdirs==1.4.4
arrow==0.12.1
asn1crypto==0.24.0
astor==0.8.0
async-generator==1.10
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.8
auto-sklearn==0.5.2
autopage==0.6.0
aws-sam-translator==1.12.0
aws-xray-sdk==2.4.2
awscli==1.18.106
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bert-tensorflow==1.0.1
black==19.10b0
blaze==0.10.1
bleach==3.0.2
blinker==1.4
blis==0.2.4
bokeh==1.0.2
boto==2.49.0
boto3==1.14.29
botocore==1.17.29
bs4==0.0.1
bunch==1.0.1
catboost==0.14.2
certifi==2021.5.30
cffi==1.11.4
cfn-lint==0.22.1
chardet==4.0.0
charset-normalizer==2.0.4
Click==7.0
cliff==3.10.1
cloudpickle==1.2.1
cmaes==0.9.0
cmd2==2.4.3
colorama==0.4.3
colorlog==6.10.1
colorlover==0.3.0
conda==4.4.10
configparser==3.7.4
ConfigSpace==0.4.10
conllu==0.11
cryptography==2.7
cufflinks==0.12.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.7
cytoolz==0.8.2
dask==2.1.0
datashape==0.5.2
DateTime==4.3
deap==1.3.0
decorator==4.4.2
defusedxml==0.6.0
Delorean==0.6.0
dill==0.3.0
docker==4.0.2
docopt==0.6.2
docutils==0.15.2
ecdsa==0.13.2
editdistance==0.5.3
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.1.0
fasttext==0.8.22
filelock==3.0.12
findspark==1.3.0
fitter==1.0.8
flaky==3.6.0
Flask==1.1.1
Flask-Cors==3.0.8
flatbuffers==1.11
ftfy==5.5.1
funcsigs==1.0.2
funcy==1.12
future==0.17.1
gast==0.2.2
gensim==3.7.3
Geohash==1.0
gevent==1.4.0
gpxpy==1.1.2
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.22.0
h5py==2.8.0rc1
haversine==0.4.5
heamy==0.0.7
hmmlearn==0.2.1
humanize==0.5.1
idna==3.2
imagesize==1.1.0
importlib-metadata==1.7.0
importlib-resources==5.4.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.5.0
itsdangerous==1.1.0
jedi==0.13.2
jieba==0.39
Jinja2==2.10
jmespath==0.10.0
joblib==0.13.2
jsondiff==1.1.2
jsonnet==0.13.0
jsonpatch==1.23
jsonpickle==1.2
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.0.0
jupyter-core==4.6.3
jupyter-kernel-gateway==1.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
klab-autotime==0.0.2
langid==1.1.6
liac-arff==2.4.0
librosa==0.6.1
lightgbm==2.2.3
line-profiler==2.1.2
llvmlite==0.29.0
lockfile==0.12.2
lxml==4.2.1
Mako==1.1.6
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.2
matplotlib-venn==0.11.5
MDP==3.5
missingno==0.4.0
mistune==0.8.4
ml-metrics==0.1.4
mlxtend==0.16.0
mock==3.0.5
modin==0.5.0
more-itertools==7.1.0
moto==1.3.9
mpld3==0.3
mplleaflet==0.0.5
mpmath==1.1.0
msgpack==0.5.6
multipledispatch==0.6.0
murmurhash==1.0.2
mwutils @ git+https://github.com.cnpmjs.org/Kesci/mwutils.git@843c5e0226f9d5498a8e27c038dbbec16e97cac9
mxnet==1.4.1
nbclient==0.4.1
nbconvert==5.6.1
nbformat==5.0.7
nest-asyncio==1.4.0
netaddr==0.7.19
networkx==2.3
nibabel==2.4.1
NiftyNet==0.3.0
nltk==3.4.1
nose==1.3.7
notebook==4.4.1
numba==0.44.1
numexpr==2.6.9
numpy==1.16.3
numpydoc==0.9.1
odo==0.5.0
onnx==1.3.0
opencv-python==4.1.0.25
openpyxl==3.0.10
optuna==3.0.6
orderedmultidict==1.0
overrides==1.9
Optuna
packaging==21.3
paddlepaddle==1.5.0
pandas==0.24.2
pandas-profiling==1.4.2
pandocfilters==1.4.2
papermill==2.1.1
parsimonious==0.8.1
parso==0.3.1
pathspec==0.8.0
patsy==0.5.1
pbr==7.0.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
plac==0.9.6
plotly==3.9.0
pluggy==0.12.0
preshed==2.0.1
prettytable==0.7.2
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.8.0
psutil==5.8.0
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pudb==2017.1
py==1.8.0
py-cpuinfo==5.0.0
py-lz4framed==0.13.0
py4j==0.10.7
pyasn1==0.4.8
pybind11==2.3.0
pycocotools==2.0.3
pycosat==0.6.3
pycparser==2.18
pydot==1.2.3
pyecharts==1.1.0
pygal==2.4.0
Pygments==2.3.1
PyJWT==2.1.0
pyLDAvis==2.1.1
pyltr==0.2.4
PyMySQL==0.10.1
pynisher==0.5.0
pynvml==11.0.0
pyOpenSSL==17.5.0
pyparsing==2.1.10
pyperclip==1.11.0
pyrfr==0.7.4
pyrsistent==0.16.0
PySocks==1.6.7
pyspark==2.4.2
pystan==2.18.0.0
pytest==5.0.1
python-dateutil==2.8.1
python-jose==3.0.1
python-Levenshtein==0.12.0
python-speech-features==0.6
pytorch-pretrained-bert==0.6.2
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.5.1
randomgen==1.16.6
rarfile==3.0
ray==0.6.6
recordio==0.1.7
redis==3.2.1
regex==2019.6.8
requests==2.26.0
resampy==0.2.1
responses==0.10.6
retrying==1.3.3
rsa==4.5
ruamel_yaml==0.15.35
s2sphere==0.2.4
s3transfer==0.3.3
sacred==0.6.10
scikit-image==0.14.2
scikit-learn==0.21.1
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
SexMachine==0.1.1
six==1.15.0
smac==0.8.0
smart-open==1.8.4
smhasher==0.150.1
snowballstemmer==1.9.0
soupsieve==1.9.2
spacy==2.1.4
Sphinx==2.1.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.5
sqlparse==0.3.0
srsly==0.0.7
sshpubkeys==3.1.0
statsmodels==0.9.0
stevedore==3.5.2
sympy==1.2
tables==3.4.2
tenacity==6.2.0
tensorboard==1.13.1
tensorboardX==1.8
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textblob==0.15.1
textwrap3==0.9.2
Theano==1.0.4
thinc==7.0.4
toml==0.10.1
toolz==0.8.2
torch==1.1.0
torchvision==0.3.0
tornado==4.5.3
TPOT==0.6.8
tqdm==4.49.0
traitlets==4.3.3
trueskill==0.4.4
typed-ast==1.4.1
typing==3.7.4
typing-extensions==3.7.4
tzlocal==1.5.1
Unidecode==1.1.1
update-checker==0.16
urllib3==1.26.6
urwid==2.0.1
vega==2.4.0
vida==0.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
widgetsnbextension==3.5.0
word2number==1.1
Wordbatch==1.3.8
wordcloud==1.5.0
wrapt==1.11.2
xgboost
xlearn==0.40a1
xlrd==1.2.0
xmltodict==0.12.0
zipp==3.1.0
zope.interface==4.6.0
